변수, 함수 지정자

+++ basic/basic_2.sol
@@ -26,30 +26,30 @@ public, external, private, internal
 */
 
 //SPDX-License-Identifier:GPL-3.0
-pragma solidity >= 0.7.0 < 0.9.0;
+pragma solidity >=0.7.0 <0.9.0;
 
-contract basic_2{
+contract basic_2 {
     //  매개 변수를 받지 않으면서 출력값도 없는 기본 함수
-    uint public a = 7;
+    uint256 public a = 7;
 
-    function example_1() public{
-        a = 14;        
+    function example_1() public {
+        a = 14;
     }
 
     // 매개변수를 받는 함수
-    function example_2(uint256 _num) public{
+    function example_2(uint256 _num) public {
         a = _num;
     }
 
     // 두 개 이상의 매개변수를 가진 함수
-    function example_3(uint256 _num1, uint256 _num2) public{   
+    function example_3(uint256 _num1, uint256 _num2) public {
         uint256 b = 2;
         a = _num1;
         b = _num2;
     }
 
     // 값을 반환하는 함수
-    function example_4(uint256 _num1) public returns(uint256){
+    function example_4(uint256 _num1) public returns (uint256) {
         a = _num1;
         return a;
     }
